backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
Pillow==10.0.0
shap==0.42.1
Werkzeug==2.3.7
